Add Truncate functionality
The new file is created with the amount of bytes from the previous
version. The file is closed and opened again several times as siva files
can not be opened in read and write modes at the same time.

The last step opens the file again with the same flags as the original one
and copies the structure.

A new test is added for Truncate as the one implemented in go-billy uses
a file in RDWD mode.

Signed-off-by: Javi Fontan <jfontan@gmail.com>